Dodana konfiguracja do połączenia do bazy.
