+++ back-end/.dockerignore
@@ -1,2 +1,3 @@
 bin\
 obj\
+.vs\
